requirements.txt
hermes-python>=0.1
pygame>=1.9.6
# pygame-menu>=2.0.4
pandas>=0.25
pillow>=6.1